requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
bidict==0.23.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
docutils==0.21.2
fastapi==0.115.4
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
markdown2==2.5.1
MarkupSafe==3.0.2
multidict==6.1.0
nicegui==2.5.0
numpy==2.1.3
opencv-python==4.10.0.84
orjson==3.10.11
propcache==0.2.0
pscript==0.7.7
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dotenv==1.0.1
python-engineio==4.10.1
python-multipart==0.0.17
python-socketio==5.11.4
PyYAML==6.0.2
requests==2.32.3
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
vbuild==0.8.2
watchfiles==0.24.0
websockets==13.1
wsproto==1.2.0
yarl==1.17.1
